// web/src/api/generated/model/alertRuleResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Condux Control Plane API
 * OpenAPI spec version: v1
 */
import type { AlertChannelResponse } from './alertChannelResponse';

export interface AlertRuleResponse {
  id: string;
  name: string;
  events: number[];
  levels: number[];
  enabled: boolean;
  channels: AlertChannelResponse[];
}